pyproject.toml
[project]
name = "pydantic-agent"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "pydantic-ai-slim[anthropic]",
    "ddgs",
    "httpx",
    "trafilatura",
]
